# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for bus-api.",
        "title": "bus-api API",
        "contact": {
            "name": "Lewis",
            "email": "lxx0103@yahoo.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/wxusers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序用户管理"
                ],
                "summary": "小程序用户列表",
                "operationId": "301",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数（5/10/15/20）",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "班级",
                        "name": "class",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "年级",
                        "name": "grade",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "身份证号",
                        "name": "identity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色（学生，员工）",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/auth.WxUserResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序用户管理"
                ],
                "summary": "新建小程序用户",
                "operationId": "302",
                "parameters": [
                    {
                        "description": "小程序用户信息",
                        "name": "wxUser_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.WxUserNew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.WxUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/wxusers/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序用户管理"
                ],
                "summary": "根据ID获取小程序用户",
                "operationId": "304",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小程序用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.WxUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序用户管理"
                ],
                "summary": "根据ID更新小程序用户",
                "operationId": "303",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小程序用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "小程序用户信息",
                        "name": "wxUser_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.WxUserNew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.WxUser"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序用户管理"
                ],
                "summary": "根据ID删除小程序用户",
                "operationId": "305",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小程序用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/wxusers/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序用户管理"
                ],
                "summary": "批量导入微信用户",
                "operationId": "61",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "auth.WxUserResponse": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "expire_date": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identity": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "open_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ListRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "page_id": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "response.SuccessRes": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "user.WxUser": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "wxUser_id": {
                    "type": "string"
                },
                "wxUser_type": {
                    "type": "string"
                }
            }
        },
        "user.WxUserNew": {
            "type": "object",
            "required": [
                "expire_date",
                "identity",
                "name",
                "role",
                "status"
            ],
            "properties": {
                "class": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "expire_date": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "grade": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "identity": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "role": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "user.WxUserResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "wxUser_id": {
                    "type": "string"
                }
            }
        }
    }
}